+++ ImTK/UI/IRenderRoot.cs
@@ -2,6 +2,5 @@ namespace ImTK.UI
 {
     public interface IRenderRoot
     {
-        RenderListCache RenderCache { get; }
     }
 }